+++ src/libpcap/__about__.py
@@ -1,6 +1,6 @@
-# Copyright (c) 2016-2022, Adam Karpierz
+# Copyright (c) 2016 Adam Karpierz
 # Licensed under the BSD license
-# https://opensource.org/licenses/BSD-3-Clause
+# https://opensource.org/license/bsd-3-clause
 
 __import__("pkg_about").about()
-__copyright__ = f"Copyright (c) 2016-2022, {__author__}"  # noqa
+__copyright__ = f"Copyright (c) 2016-2024 {__author__}"  # noqa