+++ src/libpcap/pcap.h/pcap.h
@@ -127,7 +127,7 @@
   #include <sys/time.h>
 #endif /* _WIN32/MSDOS/UN*X */
 
-#include <pcap/socket.h>	/* for SOCKET, as the active-mode rpcap APIs use it */
+#include <pcap/socket.h>	/* for PCAP_SOCKET, as the active-mode rpcap APIs use it */
 
 #ifndef PCAP_DONT_INCLUDE_PCAP_BPF_H
 #include <pcap/bpf.h>
@@ -208,8 +208,8 @@ struct pcap_file_header {
 	bpf_u_int32 magic;
 	u_short version_major;
 	u_short version_minor;
-	bpf_int32 thiszone;	/* gmt to local correction; this is always 0 */
-	bpf_u_int32 sigfigs;	/* accuracy of timestamps; this is always 0 */
+	bpf_int32 thiszone;	/* not used - SHOULD be filled with 0 */
+	bpf_u_int32 sigfigs;	/* not used - SHOULD be filled with 0 */
 	bpf_u_int32 snaplen;	/* max length saved portion of each pkt */
 	bpf_u_int32 linktype;	/* data link type (LINKTYPE_*) */
 };
@@ -310,39 +310,6 @@ struct pcap_stat {
 #endif /* _WIN32 */
 };
 
-#ifdef MSDOS
-/*
- * As returned by the pcap_stats_ex()
- */
-struct pcap_stat_ex {
-       u_long  rx_packets;        /* total packets received       */
-       u_long  tx_packets;        /* total packets transmitted    */
-       u_long  rx_bytes;          /* total bytes received         */
-       u_long  tx_bytes;          /* total bytes transmitted      */
-       u_long  rx_errors;         /* bad packets received         */
-       u_long  tx_errors;         /* packet transmit problems     */
-       u_long  rx_dropped;        /* no space in Rx buffers       */
-       u_long  tx_dropped;        /* no space available for Tx    */
-       u_long  multicast;         /* multicast packets received   */
-       u_long  collisions;
-
-       /* detailed rx_errors: */
-       u_long  rx_length_errors;
-       u_long  rx_over_errors;    /* receiver ring buff overflow  */
-       u_long  rx_crc_errors;     /* recv'd pkt with crc error    */
-       u_long  rx_frame_errors;   /* recv'd frame alignment error */
-       u_long  rx_fifo_errors;    /* recv'r fifo overrun          */
-       u_long  rx_missed_errors;  /* recv'r missed packet         */
-
-       /* detailed tx_errors */
-       u_long  tx_aborted_errors;
-       u_long  tx_carrier_errors;
-       u_long  tx_fifo_errors;
-       u_long  tx_heartbeat_errors;
-       u_long  tx_window_errors;
-     };
-#endif
-
 /*
  * Item in a list of interfaces.
  */
@@ -429,9 +396,11 @@ typedef void (*pcap_handler)(u_char *, const struct pcap_pkthdr *,
  */
 #define PCAP_CHAR_ENC_LOCAL	0x00000000U	/* strings are in the local character encoding */
 #define PCAP_CHAR_ENC_UTF_8	0x00000001U	/* strings are in UTF-8 */
+#define PCAP_MMAP_32BIT	0x00000002U	/* map packet buffers with 32-bit addresses */
 
 PCAP_AVAILABLE_1_10
-PCAP_API int	pcap_init(unsigned int, char *);
+PCAP_API int	pcap_init(unsigned int, char *)
+	    PCAP_NONNULL(2) PCAP_WARN_UNUSED_RESULT;
 
 /*
  * We're deprecating pcap_lookupdev() for various reasons (not
@@ -443,43 +412,45 @@ PCAP_DEPRECATED("use 'pcap_findalldevs' and use the first device")
 PCAP_API char	*pcap_lookupdev(char *);
 
 PCAP_AVAILABLE_0_4
-PCAP_API int	pcap_lookupnet(const char *, bpf_u_int32 *, bpf_u_int32 *, char *);
+PCAP_API int	pcap_lookupnet(const char *, bpf_u_int32 *, bpf_u_int32 *,
+	    char *) PCAP_NONNULL(4) PCAP_WARN_UNUSED_RESULT;
 
 PCAP_AVAILABLE_1_0
-PCAP_API pcap_t	*pcap_create(const char *, char *);
+PCAP_API pcap_t	*pcap_create(const char *, char *) PCAP_NONNULL(2);
 
 PCAP_AVAILABLE_1_0
-PCAP_API int	pcap_set_snaplen(pcap_t *, int);
+PCAP_API int	pcap_set_snaplen(pcap_t *, int) PCAP_WARN_UNUSED_RESULT;
 
 PCAP_AVAILABLE_1_0
-PCAP_API int	pcap_set_promisc(pcap_t *, int);
+PCAP_API int	pcap_set_promisc(pcap_t *, int) PCAP_WARN_UNUSED_RESULT;
 
 PCAP_AVAILABLE_1_0
 PCAP_API int	pcap_can_set_rfmon(pcap_t *);
 
 PCAP_AVAILABLE_1_0
-PCAP_API int	pcap_set_rfmon(pcap_t *, int);
+PCAP_API int	pcap_set_rfmon(pcap_t *, int) PCAP_WARN_UNUSED_RESULT;
 
 PCAP_AVAILABLE_1_0
-PCAP_API int	pcap_set_timeout(pcap_t *, int);
+PCAP_API int	pcap_set_timeout(pcap_t *, int) PCAP_WARN_UNUSED_RESULT;
 
 PCAP_AVAILABLE_1_2
-PCAP_API int	pcap_set_tstamp_type(pcap_t *, int);
+PCAP_API int	pcap_set_tstamp_type(pcap_t *, int) PCAP_WARN_UNUSED_RESULT;
 
 PCAP_AVAILABLE_1_5
-PCAP_API int	pcap_set_immediate_mode(pcap_t *, int);
+PCAP_API int	pcap_set_immediate_mode(pcap_t *, int) PCAP_WARN_UNUSED_RESULT;
 
 PCAP_AVAILABLE_1_0
-PCAP_API int	pcap_set_buffer_size(pcap_t *, int);
+PCAP_API int	pcap_set_buffer_size(pcap_t *, int) PCAP_WARN_UNUSED_RESULT;
 
 PCAP_AVAILABLE_1_5
-PCAP_API int	pcap_set_tstamp_precision(pcap_t *, int);
+PCAP_API int	pcap_set_tstamp_precision(pcap_t *, int)
+	    PCAP_WARN_UNUSED_RESULT;
 
 PCAP_AVAILABLE_1_5
-PCAP_API int	pcap_get_tstamp_precision(pcap_t *);
+PCAP_API int	pcap_get_tstamp_precision(pcap_t *) PCAP_WARN_UNUSED_RESULT;
 
 PCAP_AVAILABLE_1_0
-PCAP_API int	pcap_activate(pcap_t *);
+PCAP_API int	pcap_activate(pcap_t *) PCAP_WARN_UNUSED_RESULT;
 
 PCAP_AVAILABLE_1_2
 PCAP_API int	pcap_list_tstamp_types(pcap_t *, int **);
@@ -498,7 +469,7 @@ PCAP_API const char *pcap_tstamp_type_val_to_description(int);
 
 #ifdef __linux__
 PCAP_AVAILABLE_1_9
-PCAP_API int	pcap_set_protocol_linux(pcap_t *, int);
+PCAP_API int	pcap_set_protocol_linux(pcap_t *, int) PCAP_WARN_UNUSED_RESULT;
 #endif
 
 /*
@@ -560,7 +531,8 @@ PCAP_API int	pcap_set_protocol_linux(pcap_t *, int);
 #define PCAP_TSTAMP_PRECISION_NANO	1	/* use timestamps with nanosecond precision */
 
 PCAP_AVAILABLE_0_4
-PCAP_API pcap_t	*pcap_open_live(const char *, int, int, int, char *);
+PCAP_API pcap_t	*pcap_open_live(const char *, int, int, int, char *)
+	    PCAP_NONNULL(5);
 
 PCAP_AVAILABLE_0_6
 PCAP_API pcap_t	*pcap_open_dead(int, int);
@@ -569,16 +541,18 @@ PCAP_AVAILABLE_1_5
 PCAP_API pcap_t	*pcap_open_dead_with_tstamp_precision(int, int, u_int);
 
 PCAP_AVAILABLE_1_5
-PCAP_API pcap_t	*pcap_open_offline_with_tstamp_precision(const char *, u_int, char *);
+PCAP_API pcap_t	*pcap_open_offline_with_tstamp_precision(const char *, u_int,
+	    char *) PCAP_NONNULL(3);
 
 PCAP_AVAILABLE_0_4
-PCAP_API pcap_t	*pcap_open_offline(const char *, char *);
+PCAP_API pcap_t	*pcap_open_offline(const char *, char *) PCAP_NONNULL(2);
 
 #ifdef _WIN32
   PCAP_AVAILABLE_1_5
-  PCAP_API pcap_t  *pcap_hopen_offline_with_tstamp_precision(intptr_t, u_int, char *);
+  PCAP_API pcap_t  *pcap_hopen_offline_with_tstamp_precision(intptr_t, u_int,
+	    char *) PCAP_NONNULL(3);
 
-  PCAP_API pcap_t  *pcap_hopen_offline(intptr_t, char *);
+  PCAP_API pcap_t  *pcap_hopen_offline(intptr_t, char *) PCAP_NONNULL(2);
   /*
    * If we're building libpcap, these are internal routines in savefile.c,
    * so we must not define them as macros.
@@ -600,50 +574,59 @@ PCAP_API pcap_t	*pcap_open_offline(const char *, char *);
   #endif
 #else /*_WIN32*/
   PCAP_AVAILABLE_1_5
-  PCAP_API pcap_t	*pcap_fopen_offline_with_tstamp_precision(FILE *, u_int, char *);
+  PCAP_API pcap_t	*pcap_fopen_offline_with_tstamp_precision(FILE *,
+	    u_int, char *) PCAP_NONNULL(3);
 
   PCAP_AVAILABLE_0_9
-  PCAP_API pcap_t	*pcap_fopen_offline(FILE *, char *);
+  PCAP_API pcap_t	*pcap_fopen_offline(FILE *, char *) PCAP_NONNULL(2);
 #endif /*_WIN32*/
 
 PCAP_AVAILABLE_0_4
 PCAP_API void	pcap_close(pcap_t *);
 
 PCAP_AVAILABLE_0_4
-PCAP_API int	pcap_loop(pcap_t *, int, pcap_handler, u_char *);
+PCAP_API int	pcap_loop(pcap_t *, int, pcap_handler, u_char *)
+	    PCAP_WARN_UNUSED_RESULT;
 
 PCAP_AVAILABLE_0_4
-PCAP_API int	pcap_dispatch(pcap_t *, int, pcap_handler, u_char *);
+PCAP_API int	pcap_dispatch(pcap_t *, int, pcap_handler, u_char *)
+	    PCAP_WARN_UNUSED_RESULT;
 
 PCAP_AVAILABLE_0_4
 PCAP_API const u_char *pcap_next(pcap_t *, struct pcap_pkthdr *);
 
 PCAP_AVAILABLE_0_8
-PCAP_API int 	pcap_next_ex(pcap_t *, struct pcap_pkthdr **, const u_char **);
+PCAP_API int	pcap_next_ex(pcap_t *, struct pcap_pkthdr **, const u_char **);
 
 PCAP_AVAILABLE_0_8
 PCAP_API void	pcap_breakloop(pcap_t *);
 
 PCAP_AVAILABLE_0_4
-PCAP_API int	pcap_stats(pcap_t *, struct pcap_stat *);
+PCAP_API int	pcap_stats(pcap_t *, struct pcap_stat *)
+	    PCAP_WARN_UNUSED_RESULT;
 
 PCAP_AVAILABLE_0_4
-PCAP_API int	pcap_setfilter(pcap_t *, struct bpf_program *);
+PCAP_API int	pcap_setfilter(pcap_t *, struct bpf_program *)
+	     PCAP_WARN_UNUSED_RESULT;
 
 PCAP_AVAILABLE_0_9
-PCAP_API int 	pcap_setdirection(pcap_t *, pcap_direction_t);
+PCAP_API int	pcap_setdirection(pcap_t *, pcap_direction_t)
+	     PCAP_WARN_UNUSED_RESULT;
 
 PCAP_AVAILABLE_0_7
 PCAP_API int	pcap_getnonblock(pcap_t *, char *);
 
 PCAP_AVAILABLE_0_7
-PCAP_API int	pcap_setnonblock(pcap_t *, int, char *);
+PCAP_API int	pcap_setnonblock(pcap_t *, int, char *)
+	     PCAP_WARN_UNUSED_RESULT;
 
 PCAP_AVAILABLE_0_9
-PCAP_API int	pcap_inject(pcap_t *, const void *, size_t);
+PCAP_API int	pcap_inject(pcap_t *, const void *, size_t)
+	    PCAP_WARN_UNUSED_RESULT;
 
 PCAP_AVAILABLE_0_8
-PCAP_API int	pcap_sendpacket(pcap_t *, const u_char *, int);
+PCAP_API int	pcap_sendpacket(pcap_t *, const u_char *, int)
+	    PCAP_WARN_UNUSED_RESULT;
 
 PCAP_AVAILABLE_1_0
 PCAP_API const char *pcap_statustostr(int);
@@ -659,12 +642,12 @@ PCAP_API void	pcap_perror(pcap_t *, const char *);
 
 PCAP_AVAILABLE_0_4
 PCAP_API int	pcap_compile(pcap_t *, struct bpf_program *, const char *, int,
-	    bpf_u_int32);
+	    bpf_u_int32) PCAP_WARN_UNUSED_RESULT;
 
 PCAP_AVAILABLE_0_5
 PCAP_DEPRECATED("use pcap_open_dead(), pcap_compile() and pcap_close()")
 PCAP_API int	pcap_compile_nopcap(int, int, struct bpf_program *,
-	    const char *, int, bpf_u_int32);
+	    const char *, int, bpf_u_int32) PCAP_WARN_UNUSED_RESULT;
 
 /* XXX - this took two arguments in 0.4 and 0.5 */
 PCAP_AVAILABLE_0_6
@@ -684,7 +667,7 @@ PCAP_AVAILABLE_0_8
 PCAP_API int	pcap_list_datalinks(pcap_t *, int **);
 
 PCAP_AVAILABLE_0_8
-PCAP_API int	pcap_set_datalink(pcap_t *, int);
+PCAP_API int	pcap_set_datalink(pcap_t *, int) PCAP_WARN_UNUSED_RESULT;
 
 PCAP_AVAILABLE_0_8
 PCAP_API void	pcap_free_datalinks(int *);
@@ -698,7 +681,7 @@ PCAP_API const char *pcap_datalink_val_to_name(int);
 PCAP_AVAILABLE_0_8
 PCAP_API const char *pcap_datalink_val_to_description(int);
 
-PCAP_AVAILABLE_1_10
+PCAP_AVAILABLE_1_9
 PCAP_API const char *pcap_datalink_val_to_description_or_dlt(int);
 
 PCAP_AVAILABLE_0_4
@@ -791,7 +774,8 @@ PCAP_AVAILABLE_0_4
 PCAP_API void	pcap_dump(u_char *, const struct pcap_pkthdr *, const u_char *);
 
 PCAP_AVAILABLE_0_7
-PCAP_API int	pcap_findalldevs(pcap_if_t **, char *);
+PCAP_API int	pcap_findalldevs(pcap_if_t **, char *)
+	    PCAP_NONNULL(2) PCAP_WARN_UNUSED_RESULT;
 
 PCAP_AVAILABLE_0_7
 PCAP_API void	pcap_freealldevs(pcap_if_t *);
@@ -841,17 +825,19 @@ PCAP_API const char *pcap_lib_version(void);
     typedef struct _AirpcapHandle *PAirpcapHandle;
   #endif
 
-  PCAP_API int pcap_setbuff(pcap_t *p, int dim);
-  PCAP_API int pcap_setmode(pcap_t *p, int mode);
-  PCAP_API int pcap_setmintocopy(pcap_t *p, int size);
+  PCAP_API int pcap_setbuff(pcap_t *p, int dim) PCAP_WARN_UNUSED_RESULT;
+  PCAP_API int pcap_setmode(pcap_t *p, int mode) PCAP_WARN_UNUSED_RESULT;
+  PCAP_API int pcap_setmintocopy(pcap_t *p, int size) PCAP_WARN_UNUSED_RESULT;
 
   PCAP_API HANDLE pcap_getevent(pcap_t *p);
 
   PCAP_AVAILABLE_1_8
-  PCAP_API int pcap_oid_get_request(pcap_t *, bpf_u_int32, void *, size_t *);
+  PCAP_API int pcap_oid_get_request(pcap_t *, bpf_u_int32, void *, size_t *)
+	     PCAP_WARN_UNUSED_RESULT;
 
   PCAP_AVAILABLE_1_8
-  PCAP_API int pcap_oid_set_request(pcap_t *, bpf_u_int32, const void *, size_t *);
+  PCAP_API int pcap_oid_set_request(pcap_t *, bpf_u_int32, const void *,
+	    size_t *) PCAP_WARN_UNUSED_RESULT;
 
   PCAP_API pcap_send_queue* pcap_sendqueue_alloc(u_int memsize);
 
@@ -863,11 +849,13 @@ PCAP_API const char *pcap_lib_version(void);
 
   PCAP_API struct pcap_stat *pcap_stats_ex(pcap_t *p, int *pcap_stat_size);
 
-  PCAP_API int pcap_setuserbuffer(pcap_t *p, int size);
+  PCAP_API int pcap_setuserbuffer(pcap_t *p, int size) PCAP_WARN_UNUSED_RESULT;
 
-  PCAP_API int pcap_live_dump(pcap_t *p, char *filename, int maxsize, int maxpacks);
+  PCAP_API int pcap_live_dump(pcap_t *p, char *filename, int maxsize,
+	    int maxpacks) PCAP_WARN_UNUSED_RESULT;
 
-  PCAP_API int pcap_live_dump_ended(pcap_t *p, int sync);
+  PCAP_API int pcap_live_dump_ended(pcap_t *p, int sync)
+	    PCAP_WARN_UNUSED_RESULT;
 
   PCAP_API int pcap_start_oem(char* err_str, int flags);
 
@@ -876,16 +864,6 @@ PCAP_API const char *pcap_lib_version(void);
   #define MODE_CAPT 0
   #define MODE_STAT 1
   #define MODE_MON 2
-
-#elif defined(MSDOS)
-
-  /*
-   * MS-DOS definitions
-   */
-
-  PCAP_API int  pcap_stats_ex (pcap_t *, struct pcap_stat_ex *);
-  PCAP_API void pcap_set_wait (pcap_t *p, void (*yield)(void), int wait);
-  PCAP_API u_long pcap_mac_packets (void);
 
 #else /* UN*X */
 
@@ -899,13 +877,34 @@ PCAP_API const char *pcap_lib_version(void);
   PCAP_AVAILABLE_1_9
   PCAP_API const struct timeval *pcap_get_required_select_timeout(pcap_t *);
 
-#endif /* _WIN32/MSDOS/UN*X */
+#endif /* _WIN32/UN*X */
 
 /*
- * Remote capture definitions.
+ * APIs.added in WinPcap for remote capture.
  *
- * These routines are only present if libpcap has been configured to
- * include remote capture support.
+ * They are present even if remote capture isn't enabled, as they
+ * also support local capture, and as their absence may complicate
+ * code build on macOS 14 with Xcode 15, as that platform supports
+ * "weakly linked symbols":
+ *
+ *    https://developer.apple.com/library/archive/documentation/MacOSX/Conceptual/BPFrameworks/Concepts/WeakLinking.html
+ *
+ * which are symbols in dynamically-linked shared libraries, declare in
+ * such a fashion that if a program linked against a newer software
+ * development kit (SDK), and using a symbol present in the OS version
+ * for which that SDK is provided, is run on an older OS version that
+ * lacks that symbol, that symbol's value is a NULL pointer.  This
+ * allows those programs to test for the presence of that symbol
+ * by checking whether it's non-null and, if it is, using the symbol,
+ * otherwise not using it.
+ *
+ * (This is a slightly more convenient alternative to the usual
+ * technique used on Windows - and also available, and sometimes
+ * used, on UN*Xes - of loading the library containing the symbol
+ * at run time with dlopen() on UN*Xes and LoadLibrary() on Windows,
+ * looking up the symbol with dlsym() on UN*Xes and GetProcAddress()
+ * on Windows, and using the symbol with the returned pointer if it's
+ * not null.)
  */
 
 /*
@@ -1134,17 +1133,20 @@ struct pcap_rmtauth
  * For opening a remote capture, pcap_open() is currently the only
  * API available.
  */
-PCAP_AVAILABLE_1_9
+PCAP_AVAILABLE_1_9_REMOTE
 PCAP_API pcap_t	*pcap_open(const char *source, int snaplen, int flags,
-	    int read_timeout, struct pcap_rmtauth *auth, char *errbuf);
+	    int read_timeout, struct pcap_rmtauth *auth, char *errbuf)
+	    PCAP_NONNULL(6);
 
-PCAP_AVAILABLE_1_9
+PCAP_AVAILABLE_1_9_REMOTE
 PCAP_API int	pcap_createsrcstr(char *source, int type, const char *host,
-	    const char *port, const char *name, char *errbuf);
+	    const char *port, const char *name, char *errbuf)
+	    PCAP_NONNULL(6) PCAP_WARN_UNUSED_RESULT;
 
-PCAP_AVAILABLE_1_9
+PCAP_AVAILABLE_1_9_REMOTE
 PCAP_API int	pcap_parsesrcstr(const char *source, int *type, char *host,
-	    char *port, char *name, char *errbuf);
+	    char *port, char *name, char *errbuf)
+	    PCAP_NONNULL(6) PCAP_WARN_UNUSED_RESULT;
 
 /*
  * This routine can scan a directory for savefiles, list local capture
@@ -1165,9 +1167,10 @@ PCAP_API int	pcap_parsesrcstr(const char *source, int *type, char *host,
  * For listing remote capture devices, pcap_findalldevs_ex() is currently
  * the only API available.
  */
-PCAP_AVAILABLE_1_9
+PCAP_AVAILABLE_1_9_REMOTE
 PCAP_API int	pcap_findalldevs_ex(const char *source,
-	    struct pcap_rmtauth *auth, pcap_if_t **alldevs, char *errbuf);
+	    struct pcap_rmtauth *auth, pcap_if_t **alldevs, char *errbuf)
+	    PCAP_NONNULL(4) PCAP_WARN_UNUSED_RESULT;
 
 /*
  * Sampling methods.
@@ -1236,7 +1239,7 @@ struct pcap_samp
 /*
  * New functions.
  */
-PCAP_AVAILABLE_1_9
+PCAP_AVAILABLE_1_9_REMOTE
 PCAP_API struct pcap_samp *pcap_setsampling(pcap_t *p);
 
 /*
@@ -1246,24 +1249,27 @@ PCAP_API struct pcap_samp *pcap_setsampling(pcap_t *p);
 /* Maximum length of an host name (needed for the RPCAP active mode) */
 #define RPCAP_HOSTLIST_SIZE 1024
 
-PCAP_AVAILABLE_1_9
-PCAP_API SOCKET	pcap_remoteact_accept(const char *address, const char *port,
-	    const char *hostlist, char *connectinghost,
-	    struct pcap_rmtauth *auth, char *errbuf);
+PCAP_AVAILABLE_1_9_REMOTE
+PCAP_API PCAP_SOCKET	pcap_remoteact_accept(const char *address,
+	    const char *port, const char *hostlist, char *connectinghost,
+	    struct pcap_rmtauth *auth, char *errbuf)
+	    PCAP_NONNULL(6);
 
-PCAP_AVAILABLE_1_10
-PCAP_API SOCKET	pcap_remoteact_accept_ex(const char *address, const char *port,
-	    const char *hostlist, char *connectinghost,
-	    struct pcap_rmtauth *auth, int uses_ssl, char *errbuf);
+PCAP_AVAILABLE_1_10_REMOTE
+PCAP_API PCAP_SOCKET	pcap_remoteact_accept_ex(const char *address,
+	    const char *port, const char *hostlist, char *connectinghost,
+	    struct pcap_rmtauth *auth, int uses_ssl, char *errbuf)
+	    PCAP_NONNULL(7);
 
-PCAP_AVAILABLE_1_9
+PCAP_AVAILABLE_1_9_REMOTE
 PCAP_API int	pcap_remoteact_list(char *hostlist, char sep, int size,
-	    char *errbuf);
+	    char *errbuf) PCAP_NONNULL(4) PCAP_WARN_UNUSED_RESULT;
 
-PCAP_AVAILABLE_1_9
-PCAP_API int	pcap_remoteact_close(const char *host, char *errbuf);
+PCAP_AVAILABLE_1_9_REMOTE
+PCAP_API int	pcap_remoteact_close(const char *host, char *errbuf)
+	    PCAP_NONNULL(2) PCAP_WARN_UNUSED_RESULT;
 
-PCAP_AVAILABLE_1_9
+PCAP_AVAILABLE_1_9_REMOTE
 PCAP_API void	pcap_remoteact_cleanup(void);
 
 enum pcap_option_name {  /* never renumber this */